requirements.txt
Flask==2.3.3
kubernetes==27.2.0
PyYAML==6.0.1
requests==2.31.0
pyOpenSSL==23.2.0